chore: separated dashboard and settings stories
